feat: basic and post client auth methods are now interchangeable
Background: Many developers struggle to understand that this is a
configuration on the client and their SDKs rarely surface this, combined
with the fact that many SDKs default to POST while the defined default
should be basic this wastes developer's time.

This makes the two client auth methods interchangeable, so if the client
is configured with basic but post is provided, it works and vice versa.